Module Uuuu.Char

val is_valid : encoding -> char -> bool

is_valid e n is true iff n is an e scalar value.

val equal : encoding -> char -> char -> bool

equal e u u' is u = u'.

val compare : encoding -> char -> char -> int

compare u u' is Pervasives.compare u u'.

val unicode : encoding -> char -> Stdlib.Uchar.t

unicode e c is the decoded Unicode of encoding character c in encoding scheme e.