Module Uuuu.String

type 'a folder = 'a -> int -> [ `Malformed of string | `Uchar of Stdlib.Uchar.t ] -> 'a

The type for character folder. The integer is the index in the string where the `Uchar or `Malformed starts.

val fold : encoding -> ?off:int -> ?len:int -> 'a folder -> 'a -> string -> 'a

fold e ?off ?len f a s is f ( ... (f (f a pos u0) j1 u1) ... ) ... ) jn un where ui, ji are characters and their start position in the e encoded substring s starting at pos and len long. The default value for pos is 0 and len is String.length s - pos.